% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.r
\name{arle}
\alias{arle}
\title{Array relative log expression}
\usage{
arle(x, log2 = TRUE, normalize = TRUE)
}
\arguments{
\item{x}{an array of two-dimensional matrices, each of which represents an
individual sample}

\item{log2}{do the values need to be $log_2$ transformed (\code{TRUE}, the default)}

\item{normalize}{should the values be quantile normalized prior to calculating the RLE}
}
\description{
The relative log expression (RLE) quantifies the extent to which each sample
in a dataset differs from a "reference" sample, which represents each probe's
median value across all samples.
}
\examples{
# microarray visualization
if (require(affydata, quietly = TRUE)) {
  data("Dilution", package = "affydata")
  x <- arle(marray(Dilution, transpose = TRUE))
}

}
\seealso{
\code{\link[affyPLM]{RLE}}

Other array.transformations: \code{\link{arank}}
}

