% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/marray.r
\docType{methods}
\name{marray}
\alias{marray}
\alias{marray,AffyBatch-method}
\alias{marray,FeatureSet-method}
\alias{marray,PLMset-method}
\alias{marray,oligoPLM-method}
\title{Microarray array}
\usage{
marray(object, probes = NULL, select = NULL, transpose = FALSE, ...)

\S4method{marray}{AffyBatch}(object, probes = NULL, select = NULL,
  transpose = FALSE)

\S4method{marray}{PLMset}(object, probes = NULL, select = NULL,
  transpose = FALSE, type = "resid")

\S4method{marray}{FeatureSet}(object, probes = NULL, select = NULL,
  transpose = FALSE)

\S4method{marray}{oligoPLM}(object, probes = NULL, select = NULL,
  transpose = FALSE, type = "residuals")
}
\arguments{
\item{object}{\code{AffyBatch} or \code{PLMset} object}

\item{probes}{probe type see section for more information}

\item{select}{a numeric, character or logical vector indicating samples to
include}

\item{transpose}{\code{TRUE} (the default), ensures the reconstructed
microarrays are vertically oriented, as is typically expected. Set to
\code{FALSE} to return an array in the orientation strictly specified by
the platform coordinates}

\item{...}{additional arguments}

\item{type}{for objects containing probe-level models (e.g., \code{PLMsets}),
specify either \code{"residuals"} (the default) or \code{"weights"}}
}
\description{
Convert S4 microarray data structures into a three-dimensional array of
matrices, where each matrix corresponds to an individual sample's microarray
with values arranged to reflect the physical position of the corresponding
feature (i.e., probe) on the microarray surface.
}

