% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r, R/mimage.r
\docType{methods}
\name{mimage}
\alias{mimage}
\alias{mimage,AffyBatch-method}
\alias{mimage,FeatureSet-method}
\alias{mimage,PLMset-method}
\alias{mimage,array-method}
\alias{mimage,oligoPLM-method}
\title{Microarray image}
\usage{
mimage(object, colors, select = NULL, legend.label, nrow = NULL,
  ncol = NULL, fixed = FALSE, range = NULL, empty.rows = "fill",
  empty.thresh = 0.6, transform, ...)

\S4method{mimage}{AffyBatch}(object, colors, select = NULL, legend.label,
  nrow = NULL, ncol = NULL, fixed = FALSE, range = NULL,
  empty.rows = "fill", empty.thresh = 0.6, transform, probes)

\S4method{mimage}{PLMset}(object, colors, select = NULL, legend.label,
  nrow = NULL, ncol = NULL, fixed = FALSE, range = NULL,
  empty.rows = "fill", empty.thresh = 0.6, transform = identity, probes,
  type)

\S4method{mimage}{FeatureSet}(object, colors, select = NULL, legend.label,
  nrow = NULL, ncol = NULL, fixed = FALSE, range = NULL,
  empty.rows = "fill", empty.thresh = 0.6, transform, probes)

\S4method{mimage}{oligoPLM}(object, colors, select = NULL, legend.label,
  nrow = NULL, ncol = NULL, fixed = FALSE, range = NULL,
  empty.rows = "fill", empty.thresh = 0.6, transform, probes, type)

\S4method{mimage}{array}(object, colors, select = NULL,
  legend.label = "Values", nrow = NULL, ncol = NULL, fixed = FALSE,
  range = NULL, transform)
}
\arguments{
\item{object}{\code{AffyBatch} or \code{PLMset} object}

\item{colors}{a vector of colors used to represent probe values}

\item{select}{a numeric, character or logical vector indicating samples to
include}

\item{legend.label}{Legend label}

\item{nrow}{optional, number of rows in grid layout}

\item{ncol}{optional, number of columns in grid layout}

\item{fixed}{Force images to assume a fixed aspect ratio corresponding to their physical dimensions}

\item{range}{optional, if defined, values will be limited to the defined range}

\item{empty.rows}{Should empty rows be filled with values from neighboring
rows (the default, \code{"fill"}), or should they be dropped
(\code{"drop"}), or should they be left alone (\code{"nothing"})}

\item{empty.thresh}{what proportion of features must be missing from a row to
consider that row empty}

\item{transform}{a function to be applied to the values prior to visualizatio}

\item{...}{additional arguments}

\item{probes}{probe type see section for more information}

\item{type}{for objects containing probe-level models (e.g., \code{PLMsets}),
specify either \code{"residuals"} (the default) or \code{"weights"}}
}
\description{
Display a grid of color images representing microarray chips.
}
\details{
Visualize microarray probe intensities arranged by their physical location on
the array. A false color image is produced for each sample in the
microarray object and arranged in a grid.
}

